[lore-hook-actions] merge action blueprints into actions hook
