+++ examples/auth/package.json
@@ -17,7 +17,6 @@
     "lodash": "^4.0.0",
     "lore": "~0.9.0",
     "lore-auth": "~0.9.0",
-    "lore-hook-action-blueprints": "~0.10.0",
     "lore-hook-actions": "~0.10.0",
     "lore-hook-collections": "~0.10.0",
     "lore-hook-connect": "~0.10.0",
