--- examples/dialogs-material-ui/config/actionBlueprints.js
@@ -1,15 +1,0 @@
-/**
- * Configuration file for action blueprints
- */
-
-module.exports = {
-
-  /****************************************************************************
-  *                                                                           *
-  * No options currently exist, however this will eventually allow you to     *
-  * specify whether blueprints should be on or off, and which actions should  *
-  * have them all, which should only have some, and which should have none    *
-  *                                                                           *
-  ****************************************************************************/
-
-};
