+++ examples/dialogs-querying/index.js
@@ -7,7 +7,6 @@ window.lore = Lore;
 // Summon the app!
 Lore.summon({
   hooks: {
-    'action-blueprints': require("lore-hook-action-blueprints"),
     actions: require("lore-hook-actions"),
     collections: require("lore-hook-collections"),
     connect: require("lore-hook-connect"),
