+++ examples/infinite-scrolling/package.json
@@ -15,7 +15,6 @@
     "jquery": "2.1.4",
     "lodash": "3.10.1",
     "lore": "~0.9.0",
-    "lore-hook-action-blueprints": "~0.10.0",
     "lore-hook-actions": "~0.10.0",
     "lore-hook-collections": "~0.10.0",
     "lore-hook-connect": "~0.10.0",
