--- examples/pagination/config/actionBlueprints.js
@@ -1,17 +1,0 @@
-/**
- * Configuration file for action blueprints
- *
- * This file is where you define overrides for the default action blueprints.
- */
-
-module.exports = {
-
-  /****************************************************************************
-  *                                                                           *
-  * No options currently exist, however this will eventually allow you to     *
-  * specify whether blueprints should be on or off, and which actions should  *
-  * have them all, which should only have some, and which should have none    *
-  *                                                                           *
-  ****************************************************************************/
-
-};
