+++ examples/todomvc/package.json
@@ -22,7 +22,6 @@
     "json-server": "^0.8.16",
     "lodash": "^4.0.0",
     "lore": "~0.9.0",
-    "lore-hook-action-blueprints": "~0.10.0",
     "lore-hook-actions": "~0.10.0",
     "lore-hook-collections": "~0.10.0",
     "lore-hook-connect": "~0.10.0",
