--- packages/lore-hook-action-blueprints/README.md
@@ -1,3 +1,0 @@
-# lore-hook-action-blueprints
-
-A [Lore](http://www.lorejs.org) hook.
