--- packages/lore-hook-action-blueprints/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "lore-hook-action-blueprints",
-  "version": "0.10.0",
-  "license": "MIT",
-  "main": "./lib/index.js",
-  "description": "A lore hook that generates action blueprints",
-  "keywords": [
-    "lore",
-    "hook",
-    "actions",
-    "blueprints"
-  ],
-  "scripts": {
-    "build": "../../node_modules/.bin/babel src --out-dir lib",
-    "clean": "rimraf lib",
-    "debug": "mocha debug --compilers js:babel-core/register --recursive",
-    "prepublish": "npm run build",
-    "test": "mocha --compilers js:babel-core/register test/bootstrap.js test/**/*.spec.js --recursive"
-  },
-  "dependencies": {
-    "lodash": "^4.0.0",
-    "lore-utils": "^0.10.0"
-  },
-  "devDependencies": {
-    "chai": "3.4.1",
-    "mocha": "2.3.4",
-    "rimraf": "2.5.2"
-  }
-}
