+++ packages/lore-hook-actions/src/blueprints/destroy.js
