+++ packages/lore-hook-actions/src/index.js
@@ -4,6 +4,14 @@ var bindActionCreators = redux.bindActionCreators;
 var actionBlueprints = require('lore-actions').blueprints;
 var utils = require('lore-actions').utils;
 
+var blueprints = {
+  create: require('./blueprints/create'),
+  destroy: require('./blueprints/destroy'),
+  get: require('./blueprints/get'),
+  find: require('./blueprints/find'),
+  update: require('./blueprints/update')
+};
+
 function bindAction(action, store) {
   // if the module isn't a function, then it's an object describing the action
   // config and needs to be converted to a real function
@@ -30,11 +38,29 @@ function bindActionsToActionCreators(actions, store) {
 
 module.exports = {
 
-  dependencies: ['redux', 'action-blueprints'],
+  dependencies: ['models', 'collections', 'redux'],
 
   load: function(lore) {
+    lore.actions = lore.actions || {};
+    var models = lore.models;
+    var collections = lore.collections;
     var store = lore.store;
     var actions = lore.loader.loadActions();
+
+    // todo: should actions be created for files in /collections
+    // that have no corresponding model in /models Currently
+    // this only creates 'find' actions for things in /models
+
+    Object.keys(models).forEach(function(modelName) {
+      lore.actions[modelName] = lore.actions[modelName] || {};
+      _.assign(lore.actions[modelName], {
+        create: blueprints.create(modelName, models),
+        destroy: blueprints.destroy(modelName, models),
+        get: blueprints.get(modelName, models),
+        find: blueprints.find(modelName, collections),
+        update: blueprints.update(modelName, models)
+      })
+    });
 
     // overwrite any blueprints with custom implementations
     actions = _.defaultsDeep({}, actions, lore.actions);
