+++ packages/lore/test/defaultHooks.js
@@ -1,5 +1,4 @@
 module.exports = {
-  'action-blueprints': require("lore-hook-action-blueprints"),
   actions: require("lore-hook-actions"),
   collections: require("lore-hook-collections"),
   connect: require("lore-hook-connect"),
